fix(slack): move SlackAdapterConfig to ./types and drop Partial wrapper from createSlackAdapter
